+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -41,8 +41,7 @@
    <!-- Controller 추가 -->
    <context:component-scan base-package="com.project.myapp.member.controller" />
    <context:component-scan base-package="com.project.myapp"/> 
-   <context:component-scan base-package="com.project.myapp.test"/>  
-   
+  
    <!-- 로그인을 위한 인터셉터 -->
    <!-- 로그인을 해야 들어갈 수 있는 페이지 추가 -->
 <!--    <interceptors>
