+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -12,11 +12,11 @@
 		http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-4.3.xsd">
 	
  	
- 	<!-- Dao 추가 -->
+ 	<!-- Member Dao 추가 -->
 	<mybatis-spring:scan base-package="com.project.myapp.member.dao" />
 	
-	<!-- Service 추가 -->  	
- 	<context:component-scan base-package="com.example.myapp.member.service" /> 
+	<!-- Member Service 추가 -->  	
+ 	<context:component-scan base-package="com.project.myapp.member.service" /> 
 	
 	<!-- Flight DAO -->
 	<mybatis-spring:scan base-package="com.project.myapp.flight.dao"/>
