+++ src/main/webapp/WEB-INF/views/home.jsp
@@ -630,7 +630,7 @@
     	<jsp:include page="/WEB-INF/views/footer.jsp" />
     
     <!-- jquery plugins here-->
-    <script src="resources/js/jquery-1.12.1.min.js"></script>
+    <script src="/resources/js/jquery-1.12.1.min.js"></script>
     <!-- popper js -->
     <script src="resources/js/popper.min.js"></script>
     <!-- bootstrap js -->
